refactor: client side fetching for pool transactions and metadata
