+++ apps/hubble-stats/containers/PoolTransactionsTableContainer/tabs.ts
@@ -1,0 +1,4 @@
+export const allTab = 'All Transactions' as const;
+export const depositsTab = 'Deposits' as const;
+export const transfersTab = 'Transfers' as const;
+export const withdrawalsTab = 'Withdrawals' as const;
